pyproject.toml
[project]
name = "pktai"
version = "0.1.0"
description = "AI-assisted packet analysis in your terminal 🚀🤖📦💻"
authors = [
  { name = "Viswa Kumar" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "textual>=0.70.0",
  "pyshark>=0.6.0",
  "textual-fspicker>=0.4.0",
  "openai>=1.30.0",
  "PyYAML>=6.0.0",
]
license = { file = "LICENSE" }

[project.urls]
Homepage = "https://github.com/kspviswa/pktai"
Repository = "https://github.com/kspviswa/pktai"

[project.scripts]
pktai = "pktai_tui.app:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# uv will use this pyproject for dependency resolution

[tool.hatch.build.targets.wheel]
packages = ["src/pktai_tui"]
